Properly handle $no_proxy/localhost in a user's environment.
When a user has $http(s)_proxy defined, vmc will often fail to target a local
VCAP environment. The cause of this was that vmc was looking directly to the
http_proxy or https_proxy environment variables without checking that the host
was not in $no_proxy, nor checking whether the target resolved to localhost.
Both of these checks are handled via the open-uri module's find_proxy() method.